refactor: Move base to theme
