+++ app/src/main/java/com/android/quizcafe/main/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.android.quizcafe.core.designsystem.QuizCafeTheme
+import com.android.quizcafe.core.designsystem.theme.QuizCafeTheme
 import com.android.quizcafe.main.ui.QuizCafeApp
 
 class MainActivity : ComponentActivity() {
